// src/types/input-types.ts
export enum InputTypes {
  DATE_INPUT = 'date',
  SELECT_INPUT = 'select',
  TEXTAREA_INPUT = 'textarea',
  NUMBER_INPUT = 'number',
  TEXT_INPUT = 'text',
  PHONE_INPUT = 'phone',
  PASSWORD_INPUT = 'password',
  CHECKBOX_INPUT = 'checkbox',
  RADIO_INPUT = 'radio',
  EMAIL_INPUT = 'email',
  HIDDEN_INPUT = 'hidden',
  FILE_INPUT = 'file',
  HTML_INPUT = 'html',
  CONTROL_GROUP = 'control_group',
  TIME_INPUT = 'time',
}